adding logging for crew members, fixed issue with tag disconnection FREAKING prisma

+++ src/server/api/routers/crewMembers.ts
@@ -6,6 +6,8 @@ import { Redis } from "@upstash/redis";
 import { TRPCError } from "@trpc/server";
 
 import isMobilePhone from "validator/lib/isMobilePhone";
+import { clerkClient } from "@clerk/nextjs";
+import { id } from "lib0/function";
 
 const redis = new Redis({
   url: "https://us1-merry-snake-32728.upstash.io",
@@ -212,16 +214,33 @@ export const crewMembersRouter = createTRPCRouter({
         });
       }
 
-      const tagsToDisconnect = await ctx.prisma.project
-        .findUnique({
-          where: {
-            id: input.crewMemberId,
-          },
-        })
-        .tags();
+      const user = await clerkClient.users.getUser(authorId);
 
-      const disconnectTags = tagsToDisconnect || [];
+      const email = user?.emailAddresses[0]?.emailAddress;
 
+      if (!user || !email || user.banned) {
+        throw new TRPCError({
+          code: "UNAUTHORIZED",
+          message: "You are not authorized to perform this action",
+        });
+      }
+
+      const oldCrewMember = await ctx.prisma.crewMember.findUnique({
+        where: {
+          id: input.crewMemberId,
+        },
+        include: {
+          tags: true,
+        }
+      });
+
+      const crewData = {
+        ...oldCrewMember
+      }
+
+      const disconnectTags = oldCrewMember?.tags?.filter((tag) => {
+        return !input.tags?.includes(tag.id);
+      });
 
       const crewMember = await ctx.prisma.crewMember.update({
         where: {
@@ -247,79 +266,220 @@ export const crewMembersRouter = createTRPCRouter({
             })),
           },
         },
+        include: {
+          tags: true,
+        }
       });
 
-      return crewMember;
-    }),
+  let updatedName = false;
+  let updatedPosition = false;
+  let updatedDescription = false;
+  let updatedPhone = false;
+  let updatedEmail = false;
+  let updatedWage = false;
+  let updatedBurden = false;
+  let updatedRating = false;
+  let updatedTags = false;
 
-  create: privateProcedure
-    .input(
-      z.object({
-        name: z.string({ required_error: "Crew member name is required." }).min(3, "A crew member's name must be at least 3 characters long.").max(255, "A crew member's name cannot be longer than 255 characters."),
-        position: z.string({ required_error: "A crew member's position is required." }).min(3, "A crew member must have a position").max(255, "The position name is too long."),
-        notes: z.string().min(0).max(255, "Crew member notes must be less than 255 characters.").optional(),
-        phone: z.string({ required_error: "Phone Number is required." }).refine(isMobilePhone, "The phone number is invalid."),
-        email: z.string({ required_error: "Email is required." }).email("The email is invalid.").max(255, "Email must be less than 255 characters."),
-        tags: z.array(z.string()),
-        wage: z.number({ required_error: "A crew member must have a wage" }).min(0, "The wage must be a positive number.").max(1000000, "The wage must be less than 1000000."),
-        burden: z.number({ required_error: "A crew member must have burden" }).min(0, "The burden must be a positive number.").max(1000000, "The burden must be less than 1000000."),
-        rating: z.number({ required_error: "A rating has not been assigned to the crew member" }).min(0, "The rating must be a value greater than or equal to zero.").max(10, "The rating must be less than or equal to 10."),
-      })
-    )
-    .mutation(async ({ ctx, input }) => {
-      const authorId = ctx.currentUser;
-
-      const { success } = await rateLimit.limit(authorId);
-
-      if (!success) {
-        throw new TRPCError({
-          code: "TOO_MANY_REQUESTS",
-          message: "You have exceeded the rate limit, try again in a minute",
-        });
+  if(crewData.name !== crewMember.name) {
+    updatedName = true;
       }
 
-      const FoundCrewMember = await ctx.prisma.crewMember.findFirst({
-        where: {
-          name: input.name.trim(),
-        },
-      });
+if (crewData.position !== crewMember.position) {
+  updatedPosition = true;
+}
 
-      if (FoundCrewMember) {
-        throw new TRPCError({
-          code: "BAD_REQUEST",
-          message: "A crew member with this name already exists.",
-        });
-      }
+if (crewData.description !== crewMember.description) {
+  updatedDescription = true;
+}
 
-      const crewMember = await ctx.prisma.crewMember.create({
-        data: {
-          authorId,
-          name: input.name.trim(),
-          position: input.position.trim(),
-          description: input.notes?.trim() || "",
+if (crewData.phone !== crewMember.phone) {
+  updatedPhone = true;
+}
 
-          phone: input.phone.trim(),
-          email: input.email.trim(),
-          skills: "",
-          rating: input.rating.toString().trim(),
-          lastReviewDate: new Date(),
 
-          wage: input.wage,
-          burden: input.burden,
-          travel: "",
-          rate: 0,
-          hours: 0,
-          total: 0,
-          tags: {
-            connect: input.tags?.map((tag) => ({
-              id: tag,
-            })),
-          },
-        },
-      });
+if (crewData.email !== crewMember.email) {
+  updatedEmail = true;
+}
 
-      return crewMember;
+if (crewData.wage !== crewMember.wage) {
+  updatedWage = true;
+}
+
+if (crewData.burden !== crewMember.burden) {
+  updatedBurden = true;
+}
+
+if (crewData.rating !== crewMember.rating) {
+  updatedRating = true;
+}
+
+if (crewData.tags?.length !== crewMember.tags?.length) {
+  updatedTags = true;
+}
+
+crewData.tags?.forEach((tag) => {
+  if (!crewMember.tags?.find((t) => t.id === tag.id)) {
+    updatedTags = true;
+  }
+});
+
+
+let result = "";
+let changes = 0;
+
+if (updatedName) {
+  result += `Name: ${oldCrewMember?.name || ""} -> ${crewMember.name}\n`;
+  changes++;
+}
+
+if (updatedPosition) {
+  result += `Position: ${oldCrewMember?.position || ""} -> ${crewMember.position}\n`;
+  changes++;
+}
+
+if (updatedDescription) {
+  result += `Notes: ${oldCrewMember?.description || ""} -> ${crewMember.description}\n`;
+  changes++;
+}
+
+if (updatedPhone) {
+  result += `Phone: ${oldCrewMember?.phone || ""} -> ${crewMember.phone}\n`;
+  changes++;
+}
+
+if (updatedEmail) {
+  result += `Email: ${oldCrewMember?.email || ""} -> ${crewMember.email}\n`;
+  changes++;
+}
+
+if (updatedWage) {
+  result += `Wage: $${oldCrewMember?.wage || ""} -> $${crewMember.wage}\n`;
+  changes++;
+}
+
+if (updatedBurden) {
+  result += `Burden: $${oldCrewMember?.burden || ""} -> $${crewMember.burden}\n`;
+  changes++;
+
+}
+
+if (updatedRating) {
+  result += `Rating: ${oldCrewMember?.rating || ""} -> ${crewMember.rating}\n`;
+  changes++;
+}
+
+if (updatedTags) {
+  result += `Tags: ${oldCrewMember?.tags?.map((tag) => tag.name).join(", ") || ""} -> ${crewMember.tags?.map((tag) => tag.name).join(", ")}`;
+  changes++;
+}
+
+await ctx.prisma.log.create({
+  data: {
+    name: `Updated Crew Member ${crewMember.name}`,
+    action: "url",
+    url: `/crew/${crewMember.id}`,
+    authorId: authorId,
+    category: "crew",
+    description: `${changes} ${changes == 1 ? "change" : "changes"}: \n${result}`,
+    severity: "moderate",
+  }
+});
+
+return crewMember;
     }),
+
+create: privateProcedure
+  .input(
+    z.object({
+      name: z.string({ required_error: "Crew member name is required." }).min(3, "A crew member's name must be at least 3 characters long.").max(255, "A crew member's name cannot be longer than 255 characters."),
+      position: z.string({ required_error: "A crew member's position is required." }).min(3, "A crew member must have a position").max(255, "The position name is too long."),
+      notes: z.string().min(0).max(255, "Crew member notes must be less than 255 characters.").optional(),
+      phone: z.string({ required_error: "Phone Number is required." }).refine(isMobilePhone, "The phone number is invalid."),
+      email: z.string({ required_error: "Email is required." }).email("The email is invalid.").max(255, "Email must be less than 255 characters."),
+      tags: z.array(z.string()),
+      wage: z.number({ required_error: "A crew member must have a wage" }).min(0, "The wage must be a positive number.").max(1000000, "The wage must be less than 1000000."),
+      burden: z.number({ required_error: "A crew member must have burden" }).min(0, "The burden must be a positive number.").max(1000000, "The burden must be less than 1000000."),
+      rating: z.number({ required_error: "A rating has not been assigned to the crew member" }).min(0, "The rating must be a value greater than or equal to zero.").max(10, "The rating must be less than or equal to 10."),
+    })
+  )
+  .mutation(async ({ ctx, input }) => {
+    const authorId = ctx.currentUser;
+
+    const { success } = await rateLimit.limit(authorId);
+
+    if (!success) {
+      throw new TRPCError({
+        code: "TOO_MANY_REQUESTS",
+        message: "You have exceeded the rate limit, try again in a minute",
+      });
+    }
+
+    const user = await clerkClient.users.getUser(authorId);
+
+    const email = user?.emailAddresses[0]?.emailAddress;
+
+    if (!user || !email || user.banned) {
+      throw new TRPCError({
+        code: "UNAUTHORIZED",
+        message: "You are not authorized to perform this action",
+      });
+    }
+
+    const FoundCrewMember = await ctx.prisma.crewMember.findFirst({
+      where: {
+        name: input.name.trim(),
+      },
+    });
+
+    if (FoundCrewMember) {
+      throw new TRPCError({
+        code: "BAD_REQUEST",
+        message: "A crew member with this name already exists.",
+      });
+    }
+
+    const crewMember = await ctx.prisma.crewMember.create({
+      data: {
+        authorId,
+        name: input.name.trim(),
+        position: input.position.trim(),
+        description: input.notes?.trim() || "",
+
+        phone: input.phone.trim(),
+        email: input.email.trim(),
+        skills: "",
+        rating: input.rating.toString().trim(),
+        lastReviewDate: new Date(),
+
+        wage: input.wage,
+        burden: input.burden,
+        travel: "",
+        rate: 0,
+        hours: 0,
+        total: 0,
+        tags: {
+          connect: input.tags?.map((tag) => ({
+            id: tag,
+          })),
+        },
+      },
+    });
+
+    await ctx.prisma.log.create({
+      data: {
+        action: "url",
+        category: "blueprint",
+        name: `Created new Crew Member \"${input.name}\"`,
+        authorId: authorId,
+        url: `/crewmember/${crewMember.id}`,
+        description: `${email} created new crew member \"${input.name}\" (\"${input.position}\")`,
+        severity: "moderate",
+      }
+    })
+
+    return crewMember;
+  }),
 
   delete: privateProcedure
     .input(z.object({ crewMemberId: z.string() }))
@@ -335,11 +495,34 @@ export const crewMembersRouter = createTRPCRouter({
         });
       }
 
+      const user = await clerkClient.users.getUser(authorId);
+
+      const email = user?.emailAddresses[0]?.emailAddress;
+
+      if (!user || !email || user.banned) {
+        throw new TRPCError({
+          code: "UNAUTHORIZED",
+          message: "You are not authorized to perform this action",
+        });
+      }
+
       const crewMember = await ctx.prisma.crewMember.delete({
         where: {
           id: input.crewMemberId,
         },
       });
+
+      await ctx.prisma.log.create({
+        data: {
+          action: "none",
+          category: "blueprint",
+          name: `Deleted Crew Member \"${crewMember.name}\"`,
+          authorId: authorId,
+          url: `/#`,
+          description: `${email} deleted crew member \"${crewMember.name}\" (\"${crewMember.position}\")`,
+          severity: "critical",
+        }
+      })
 
       return crewMember;
     }),
